deleted [clientId], updated aggregate count for events, moved apis into the make-case.ts
