+++ icv/src/api/make-cases/make-case.ts
@@ -28,7 +28,12 @@ export async function createClient(client: Client) {
 export async function getAllClients() {
     const clientsCollection = collection(db, 'clients')
     const clientsSnapshot = await getDocs(clientsCollection)
-    const clients = clientsSnapshot.docs.map((doc) => doc.data() as Client)
+
+    const clients = clientsSnapshot.docs.map((doc) => ({
+        id: doc.id, // Firebase document ID
+        ...doc.data(), // Other client fields
+    })) as Client[]
+
     return clients
 }
 
@@ -39,11 +44,7 @@ export async function getClientById(id: string) {
     return client
 }
 
-export async function updateClient(id: string, client: Partial<Client>) {
-    if (ClientSchema.safeParse(client).success === false) {
-        throw new Error('Client object is invalid')
-    }
-
-    const clientsCollection = collection(db, 'clients')
-    await updateDoc(doc(clientsCollection, id), client)
-}
+export const updateClient = async (id: string, updatedClientData: Client) => {
+    const clientRef = doc(db, "clients", id);
+    await updateDoc(clientRef, updatedClientData);
+  };
